+++ src/tokens/backgroundColor.js
@@ -67,6 +67,12 @@ const light = (core) => ({
   // Tab
   TabDivider: core.Color.Neutral400,
   TabHover: core.Color.Neutral100,
+  // Tag
+  TagError: core.Color.Error100,
+  TagNeutral: core.Color.Neutral200,
+  TagPrimary: core.Color.Primary100,
+  TagSuccess: core.Color.Success100,
+  TagWarning: core.Color.Warning100,
   // Table
   TableRowHover: core.Color.Neutral100,
   TableFooter: core.Color.Neutral200,
@@ -139,6 +145,12 @@ const dark = (core) => ({
   // Tab
   TabDivider: core.Color.Neutral600,
   TabHover: core.Color.Neutral700,
+  // Tag
+  TagError: core.Color.Error500,
+  TagNeutral: core.Color.Neutral600,
+  TagPrimary: core.Color.Primary500,
+  TagSuccess: core.Color.Success500,
+  TagWarning: core.Color.Warning300,
   // Table
   TableRowHover: core.Color.Neutral700,
   TableFooter: core.Color.Neutral700,