+++ src/tokens/textColor.js
@@ -1,4 +1,4 @@
-const light = core => ({
+const light = (core) => ({
   // Global
   Active: core.Color.Brand300,
   ActiveHover: core.Color.Brand400,
@@ -39,11 +39,17 @@ const light = core => ({
   ModalClose: core.Color.Neutral900,
   // Tab
   TabActive: core.Color.Brand300,
+  // Tag
+  TagError: core.Color.Error500,
+  TagNeutral: core.Color.Neutral700,
+  TagPrimary: core.Color.Primary400,
+  TagSuccess: core.Color.Success500,
+  TagWarning: core.Color.Warning500,
   // Tooltip
   Tooltip: core.Color.NeutralWhite,
 })
 
-const dark = core => ({
+const dark = (core) => ({
   ...light(core),
   // Global
   Active: core.Color.Brand200,
@@ -82,6 +88,12 @@ const dark = core => ({
   ModalClose: core.Color.NeutralWhite,
   // Tab
   TabActive: core.Color.Brand200,
+  // Tag
+  TagError: core.Color.NeutralWhite,
+  TagNeutral: core.Color.NeutralWhite,
+  TagPrimary: core.Color.NeutralWhite,
+  TagSuccess: core.Color.NeutralWhite,
+  TagWarning: core.Color.Neutral800,
 })
 
 export default {